Modified - refactor Classes to core
