+++ lib/core/widgets/curve_painter.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../provider/app_theme.dart';
+import '../../provider/app_theme.dart';
 
 class CurvePainter extends CustomPainter {
   @override
