+++ lib/features/customer_feature/presentation/screens/login_screen.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_spinkit/flutter_spinkit.dart';
 import 'package:provider/provider.dart';
 
-import '../../../../classes/http_exception.dart';
+import '../../../../core/connection/http_exception.dart';
 import '../../../../provider/Products.dart';
 import '../../../../provider/app_theme.dart';
 import '../../../../provider/auth.dart';
